Extract wwz stuff from top stuff
